Attemting to refactor vibe carousel
