+++ app/src/main/res/layout/activity_add_edit_vibe.xml
@@ -36,11 +36,6 @@
             android:orientation="vertical"
             android:padding="16dp" >
 
-            <fragment
-                android:id="@+id/vibe_field_fragment"
-                android:name="com.example.vybe.AddEdit.VibeFieldFragment"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content" />
 
             <ImageView
                 android:id="@+id/vibe_selector"
